I/O thread functions
Added read_file_thread and write_file_thread, that will read and write files on a separate thread.

Added thread_settings function that will wait for a thread to finish (block the code) and return the value that the thread function returns.

Added unit testing for file I/O on a separate thread.