Added reset_variables_allocated_on_the_heap_test function
Added reset_variables_allocated_on_the_heap_test function.

This function runs each function 1000 times to check if all variables allocated on the heap are reset.

On the first run, the function gets the amount of memory used for the test function. It then runs it 10 times and checks if the memory usage has gone up by 900%. if not it then runs 100 times and checks for 500% and then 1000 times and checks for 100%. These levels of checks will detect major and minor variables that are allocated on the heap and are not reset.

Added get_program_memory_usage function to get the current memory usage of the program.