Added tests for varint_decode
Added test for varint.

Added reset_variables_allocated_on_the_heap_test for varint_decode.

Upgraded update_delegates_online_status to use a struct for parameters so it can run on a separate thread, and it does not have to use a fixed database or collection name.

Updated the makefile to include the number_functions_test folder and files.