Changed vote global variables to struct
Changed global vote variables to struct, this way they are more organized and can hold the public addresses that voted and the amount of block verifiers.

Added reseting of struct in some server functions that were not reseting the struct before using it.

changed strnlen maximum size for some of the struct since their not using BUFFER_SIZE.