Changed how the start of round part 3 works
Round part 3 uses the same main node as round part 1. Becuase of this if a timeout of a mainnode or if a round change is valid in round part 3, the system will start over again at round part 1 instead of the same part of the round like it does with 1,2,4.

In round part 1 if their is a timeout of a mainnode or if a round change is valid, then it needs to write the backup node count of round part 1 when it is finished with round part 1 in the current_round_part_backup_node_round_part_3 field in the current_round collection in the database.

Any time round part 1 starts because of a timeout of a mainnode or if a round change is valid in round part 3, the system will read current_round_part_backup_node_round_part_3 instead of current_round_part_backup_node.

Any time round part 3 starts it will read current_round_part_backup_node_round_part_3 instead of current_round_part_backup_node.